Hotspot data only add command (#319)
* hotspot data only add command
* rm unused deps
* Refactor client, do_hotspot add
* Limit to iot and dataonly hotspots
* use pubkey deserializer

---------
