+++ Cargo.toml
@@ -29,7 +29,7 @@ hmac = "0.12"
 sha2 = "0.10"
 base64 = ">=0.21"
 bs58 = "0"
-reqwest = {version = "0", default-features=false, features=["rustls-tls"]}
+reqwest = {version = "0", default-features=false, features=["blocking", "rustls-tls"]}
 pbkdf2 = {version = "0", default-features=false }
 aes-gcm = "0"
 shamirsecretsharing = {version="0.1.5", features=["have_libsodium"]}
@@ -46,16 +46,19 @@ angry-purple-tiger = "0"
 clap = {version = "4", features = ["derive"]}
 helium-crypto = { version = "0.6.3", features = ["multisig", "solana"]}
 helium-proto = { git = "https://github.com/helium/proto", branch="master"}
-data-credits = { git = "https://github.com/helium/helium-program-library", branch="master" }
-lazy-distributor = { git = "https://github.com/helium/helium-program-library", branch="master" }
-helium-entity-manager = { git = "https://github.com/helium/helium-program-library", branch="master" }
-helium-sub-daos = { git = "https://github.com/helium/helium-program-library", branch="master" }
-circuit-breaker = { git = "https://github.com/helium/helium-program-library", branch="master" }
-spl-associated-token-account = "*"
-bincode = "*"
+data-credits = { git = "https://github.com/helium/helium-program-library", features = ["no-entrypoint"], rev = "130f71b" }
+lazy-distributor = { git = "https://github.com/helium/helium-program-library", features = ["no-entrypoint"], rev = "130f71b" }
+helium-entity-manager = { git = "https://github.com/helium/helium-program-library", rev = "130f71b", features = ["no-entrypoint"] }
+helium-sub-daos = { git = "https://github.com/helium/helium-program-library", features = ["no-entrypoint"], rev = "130f71b" }
+circuit-breaker = { git = "https://github.com/helium/helium-program-library", features = ["no-entrypoint"], rev = "130f71b" }
+mpl-bubblegum = { git = "https://github.com/metaplex-foundation/metaplex-program-library", branch = "master" }
+spl-associated-token-account = { version = "*", features = [ "no-entrypoint" ] }
+spl-account-compression = {version = "*", features = [ "no-entrypoint" ] }
+solana-program = "*"
+bincode = "1.3.3"
 pyth-sdk-solana = "0"
-anchor-client = "0"
-anchor-spl = "0"
+anchor-client = "0.26"
+anchor-spl = { version = "0.26.0", features = ["mint", "token"] }
 tiny-bip39 = "0"
 http = "0"
 jsonrpc = "0"
