+++ src/cmd/hotspots/add.rs
@@ -1,106 +1,90 @@
 use crate::{
-    b64,
-    cmd::*,
-    result::Result,
-    staking,
-    traits::{TxnEnvelope, TxnSign},
+    client::HotspotAssertion, cmd::*, dao::SubDao, hotspot::HotspotMode, result::Result,
+    traits::txn_envelope::TxnEnvelope,
 };
+use helium_proto::BlockchainTxnAddGatewayV1;
 
-#[derive(Debug, StructOpt)]
+#[derive(Clone, Debug, clap::Args)]
 /// Add a hotspot to the blockchain. The original transaction is created by the
 /// hotspot miner and supplied here for owner signing. Use an onboarding key to
 /// get the transaction signed by the DeWi staking server.
 pub struct Cmd {
-    /// Base64 encoded transaction to sign. If no transaction is given stdin is
-    /// read for the transaction. Note that the stdin feature only works if the
-    /// wallet password is set in the HELIUM_WALLET_PASSWORD environment
-    /// variable
-    #[structopt(name = "TRANSACTION")]
+    /// The subdao to assert the hotspot on. Only iot is currently supported.
+    #[arg(long)]
+    subdao: SubDao,
+
+    /// The mode of the hotspot to add. Only "dataonly" is currently supported.
+    #[arg(long)]
+    mode: HotspotMode,
+
+    /// Lattitude of hotspot location to assert.
+    ///
+    /// Defaults to the last asserted value. For negative values use '=', for
+    /// example: "--lat=-xx.xxxxxxx".
+    #[arg(long)]
+    lat: Option<f64>,
+
+    /// Longitude of hotspot location to assert.
+    ///
+    /// Defaults to the last asserted value. For negative values use '=', for
+    /// example: "--lon=-xx.xxxxxxx".
+    #[arg(long)]
+    lon: Option<f64>,
+
+    /// The antenna gain for the asserted hotspotin dBi, with one digit of
+    /// accuracy.
+    ///
+    /// Defaults to the last asserted value. Note that the gain is truncated to
+    /// the nearest 0.1 dBi.
+    #[arg(long)]
+    gain: Option<f64>,
+
+    /// The elevation for the asserted hotspot in meters above ground level.
+    ///
+    /// Defaults to the last assserted value. For negative values use '=', for
+    /// example: "--elevation=-xx".
+    #[arg(long)]
+    elevation: Option<i32>,
+
+    /// Base64 encoded transaction. If no transaction is given stdin is read for
+    /// the transaction.
+    ///
+    ///  Note that the stdin feature only works if the wallet password is set in
+    /// the HELIUM_WALLET_PASSWORD environment variable
     txn: Option<Transaction>,
 
-    /// The onboarding key to use if the payer of the transaction fees
-    /// is the DeWi "staking" server.
-    #[structopt(long)]
-    onboarding: Option<String>,
-
-    #[structopt(long)]
-    commit: bool,
+    /// Optional url for the ecc signature verifier.
+    #[arg(long, default_value = "https://ecc-verifier.web.helium.io")]
+    verifier_url: String,
+ 
+    /// Commit the hotspot add.
+    #[command(flatten)]
+    commit: CommitOpts,
 }
 
 impl Cmd {
-    pub async fn run(self, opts: Opts) -> Result {
+    pub fn run(&self, opts: Opts) -> Result {
+        if self.subdao != SubDao::Iot {
+            bail!("Only iot subdao is currently supported")
+        }
+        if self.mode != HotspotMode::DataOnly {
+            bail!("Only dataonly mode is currently supported")
+        }
+
         let mut txn = BlockchainTxnAddGatewayV1::from_envelope(&read_txn(&self.txn)?)?;
         let password = get_wallet_password(false)?;
-        let wallet = load_wallet(opts.files)?;
+        let wallet = load_wallet(&opts.files)?;
+        let client = new_client(&opts.url)?;
         let keypair = wallet.decrypt(password.as_bytes())?;
-        let staking_client = staking::Client::default();
-        let base_url: String = api_url(wallet.public_key.network);
-        let client = new_client(base_url.clone());
-        let pending_url = base_url + "/pending_transactions/";
-        let wallet_key = keypair.public_key();
 
-        txn.owner_signature = txn.sign(&keypair)?;
-        let envelope = match PublicKey::from_bytes(&txn.payer)? {
-            key if &key == wallet_key => {
-                txn.payer_signature = txn.owner_signature.clone();
-                Ok(txn.in_envelope())
-            }
-            _key if self.onboarding.is_some() && self.commit => {
-                // Only have staking server sign if there's an onboarding key,
-                // and we're actually going to commit
-                let onboarding_key = self.onboarding.as_ref().unwrap().replace('\"', "");
-                staking_client
-                    .sign(&onboarding_key, &txn.in_envelope())
-                    .await
-            }
-            _key => Ok(txn.in_envelope()),
-        }?;
-
-        let status = maybe_submit_txn(self.commit, &client, &envelope).await?;
-        print_txn(&txn, &status, &pending_url, opts.format)
-    }
-}
-
-fn print_txn(
-    txn: &BlockchainTxnAddGatewayV1,
-    status: &Option<PendingTxnStatus>,
-    pending_url: &str,
-    format: OutputFormat,
-) -> Result {
-    let address = PublicKey::from_bytes(&txn.gateway)?.to_string();
-    let owner = PublicKey::from_bytes(&txn.owner)?.to_string();
-    let payer = if txn.payer.is_empty() {
-        PublicKey::from_bytes(&txn.owner)?.to_string()
-    } else {
-        PublicKey::from_bytes(&txn.payer)?.to_string()
-    };
-    let status_endpoint = pending_url.to_owned() + status_str(status);
-    match format {
-        OutputFormat::Table => {
-            ptable!(
-                ["Key", "Value"],
-                ["Address", address],
-                ["Payer", payer],
-                ["Owner", owner],
-                ["Fee (DC)", txn.fee],
-                ["Staking fee (DC)", txn.staking_fee],
-                ["Hash", status_str(status)],
-                ["Status", status_endpoint]
-            );
-            print_footer(status)
+        if client.hotspot_key_to_asset(&txn.gateway).is_err() {
+            let tx =
+                client.hotspot_dataonly_issue(&self.verifier_url, &mut txn, keypair.clone())?;
+            self.commit.maybe_commit(&tx, &client)?;
         }
-        OutputFormat::Json => {
-            let table = json!({
-                "address": address,
-                "owner": owner,
-                "payer": payer,
-                "fee": txn.fee,
-                "staking fee": txn.staking_fee,
-                "hash": status_json(status),
-                "txn": b64::encode_message(&txn.in_envelope())?,
-                "status": status_endpoint
-            });
-            print_json(&table)
-        }
+        let assertion =
+            HotspotAssertion::try_from((self.lat, self.lon, self.elevation, self.gain))?;
+        let tx = client.hotspot_dataonly_onboard(&txn.gateway, assertion, keypair)?;
+        self.commit.maybe_commit(&tx, &client)
     }
 }
