+++ src/dao.rs
@@ -20,8 +20,33 @@ impl Dao {
             Self::Hnt => Token::Hnt.mint(),
         };
         let (dao_key, _) =
-            Pubkey::find_program_address(&[b"dao", &mint.to_bytes()], &helium_sub_daos::id());
+            Pubkey::find_program_address(&[b"dao", mint.as_ref()], &helium_sub_daos::id());
         dao_key
+    }
+
+    pub fn data_only_config_key(&self) -> Pubkey {
+        let (key, _) = Pubkey::find_program_address(
+            &[b"data_only_config", self.key().as_ref()],
+            &helium_entity_manager::id(),
+        );
+        key
+    }
+
+    pub fn entity_creator_key(&self) -> Pubkey {
+        let (key, _) = Pubkey::find_program_address(
+            &[b"entity_creator", self.key().as_ref()],
+            &helium_entity_manager::id(),
+        );
+        key
+    }
+
+    pub fn key_to_asset(&self, entity_key: &[u8]) -> Pubkey {
+        let hash = Sha256::digest(entity_key);
+        let (key, _) = Pubkey::find_program_address(
+            &[b"key_to_asset", self.key().as_ref(), hash.as_ref()],
+            &helium_entity_manager::id(),
+        );
+        key
     }
 }
 
@@ -50,7 +75,7 @@ impl SubDao {
     pub fn key(&self) -> Pubkey {
         let mint = self.mint();
         let (subdao_key, _) =
-            Pubkey::find_program_address(&[b"sub_dao", &mint.to_bytes()], &helium_sub_daos::id());
+            Pubkey::find_program_address(&[b"sub_dao", mint.as_ref()], &helium_sub_daos::id());
         subdao_key
     }
 
@@ -70,7 +95,7 @@ impl SubDao {
     pub fn delegated_dc_key(&self, router_key: &str) -> Pubkey {
         let hash = Sha256::digest(router_key);
         let (key, _) = Pubkey::find_program_address(
-            &[b"delegated_data_credits", &self.key().to_bytes(), &hash],
+            &[b"delegated_data_credits", self.key().as_ref(), &hash],
             &data_credits::id(),
         );
         key
@@ -78,7 +103,7 @@ impl SubDao {
 
     pub fn escrow_account_key(&self, delegated_dc_key: &Pubkey) -> Pubkey {
         let (key, _) = Pubkey::find_program_address(
-            &[b"escrow_dc_account", &delegated_dc_key.to_bytes()],
+            &[b"escrow_dc_account", delegated_dc_key.as_ref()],
             &data_credits::id(),
         );
         key
@@ -90,25 +115,29 @@ impl SubDao {
             Self::Mobile => b"MOBILE".as_ref(),
         };
         let (key, _) = Pubkey::find_program_address(
-            &[b"rewardable_entity_config", &self.key().to_bytes(), suffix],
+            &[b"rewardable_entity_config", self.key().as_ref(), suffix],
             &helium_entity_manager::id(),
         );
         key
     }
 
-    pub fn info_key(&self, entity_key: &helium_crypto::PublicKey) -> Result<Pubkey> {
+    pub fn info_key(&self, entity_key: &[u8]) -> Result<Pubkey> {
+        let hash = Sha256::digest(entity_key);
         let config_key = self.rewardable_entity_config_key();
-        let entity_decoded = bs58::decode(entity_key.to_string()).into_vec()?;
-        let hash = Sha256::digest(entity_decoded);
         let prefix = match self {
             Self::Iot => "iot_info",
             Self::Mobile => "mobile_info",
         };
         let (key, _) = Pubkey::find_program_address(
-            &[prefix.as_bytes(), &config_key.to_bytes(), &hash],
+            &[prefix.as_bytes(), config_key.as_ref(), &hash],
             &helium_entity_manager::id(),
         );
         Ok(key)
+    }
+
+    pub fn info_key_for_helium_key(&self, public_key: &helium_crypto::PublicKey) -> Result<Pubkey> {
+        let entity_key = bs58::decode(public_key.to_string()).into_vec()?;
+        self.info_key(&entity_key)
     }
 
     pub fn lazy_distributor_key(&self) -> Pubkey {
