+++ src/hotspot.rs
@@ -3,7 +3,7 @@ use angry_purple_tiger::AnimalName;
 use serde::Serialize;
 use std::collections::HashMap;
 
-#[derive(Debug, Serialize)]
+#[derive(Debug, Serialize, Clone, Copy, clap::ValueEnum, PartialEq, Eq)]
 #[serde(rename_all = "snake_case")]
 pub enum HotspotMode {
     Full,
@@ -17,6 +17,13 @@ impl From<bool> for HotspotMode {
         } else {
             Self::DataOnly
         }
+    }
+}
+
+impl std::fmt::Display for HotspotMode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let str = serde_json::to_string(self).map_err(|_| std::fmt::Error)?;
+        f.write_str(&str)
     }
 }
